requirements.txt
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
db-dtypes==1.4.2
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-bigquery==3.31.0
google-cloud-core==2.4.3
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
grpcio==1.72.0rc1
grpcio-status==1.72.0rc1
gspread==6.2.0
httplib2==0.22.0
idna==3.10
joblib==1.4.2
nltk==3.9.1
numpy==2.0.2
oauth2client==4.1.3
oauthlib==3.2.2
packaging==24.2
pandas==2.2.3
pandas-gbq==0.28.0
proto-plus==1.26.1
protobuf==6.30.2
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydata-google-auth==1.9.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.17.0
textblob==0.19.0
tqdm==4.67.1
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0